Small changes in archive.c

+++ archive.c
@@ -3,7 +3,7 @@
 
 int Archive(char *input, char * output) {
 	int outputDescriptor;
-	if ((outputDescriptor = open(output, O_WRONLY | O_CREAT | O_TRUNC, S_IWUSR | S_IRUSR | S_IRGRP | S_IROTH)) == (int) - 1) {
+	if ((outputDescriptor = open(output, O_WRONLY | O_CREAT | O_TRUNC, S_IWUSR | S_IRUSR | S_IRGRP | S_IROTH)) == -1) {
 		printf("Error opening file %s\n", output);
 		return 1;
 	}
@@ -50,7 +50,7 @@ int WriteArchive(char *dir, int outputDescriptor) {
 
 		if (S_ISREG(statbuf.st_mode)) { //если текущий элемент - файл
 			int fileDscr;
-			if ((fileDscr = open(entry->d_name, O_RDONLY)) == (int) - 1) {
+			if ((fileDscr = open(entry->d_name, O_RDONLY)) == -1) {
 				printf("Error opening file %s\n", entry->d_name);
 				continue;
 			}
